Fix logger to sent message

+++ Bot/VK.py
@@ -80,11 +80,12 @@ def SyncDB():
             for i in range(0,count):
                 query_find = "SELECT photo_id FROM likes WHERE id={};".format(i)
                 cursor.execute(query_find)
-                id = cursor.fetchall()[0][0]
-                if not id in ids:
-                    query_delete = "DELETE FROM likes WHERE photo_id={};".format(id)
-                    cursor.execute(query_delete)
-                    logger.info("Photo with id={} in database, but not in VK, deleted...".format(id))
+                if not cursor.fetchall():
+                    id = cursor.fetchall()[0][0]
+                    if not id in ids:
+                        query_delete = "DELETE FROM likes WHERE photo_id={};".format(id)
+                        cursor.execute(query_delete)
+                        logger.info("Photo with id={} in database, but not in VK, deleted...".format(id))
